% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em_mcem_logll.R
\name{Mest}
\alias{Mest}
\title{E-Step For Mean Estimate}
\usage{
Mest(theta, bl, bu, Freq)
}
\arguments{
\item{theta}{Initial theta value}

\item{bl}{Lower bound of the intervals, values in vector, starting from -inf.}

\item{bu}{Upper bound of the intervals, values in vector, ending with +inf.}

\item{Freq}{Frequency over the intervals, values in vector.}
}
\value{
Return M which are the estimates of M in E-step
}
\description{
E-Step For Mean Estimate
}
\examples{
 output2 =list()
 simdataaaa = univ_Simul(ncol_matrix=1,
                       n=50,
                       nclass = 10,
                       mean = 68,
                       sd = 1.80,
                       fr_breaks=c(62,64,66,68,70,72,74,76,78))

 output2<- EM(bl=simdataaaa$simul_data[,1,1],
               bu=simdataaaa$simul_data[,2,1],
               Freq=simdataaaa$simul_data[,3,1],
               theta_init=c(67,2),
               maxit = 1000,
               tol1=1e-3,
               tol2=1e-4)
 output2
}
