% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em_mcem_logll.R
\name{sigmaMCEM}
\alias{sigmaMCEM}
\title{E-Step estimate the parameter of sigma.}
\usage{
sigmaMCEM(data, simZZ, mupd)
}
\arguments{
\item{data}{Contingency table, matrix format of three columns, first column
lower limits, second column upper limits, and third column observed
frequencies. The current state of the parameters, vector containing
mean and sd.}

\item{mupd}{Is the updated estimates of mu.}

\item{simZ}{The matrix, which is the samples simulated over the intervals,
of the ZMCEM function.}
}
\value{
Return sigma which are the estimates of variance in E-step.
}
\description{
E-Step estimate the parameter of sigma.
}
\examples{
 simdataaaa = univ_Simul(ncol_matrix=1,
                         n=50,
                         nclass = 10,
                         mean = 68,
                         sd = 1.80,
                         fr_breaks=c(62,64,66,68,70,72,74,76,78))
 
 ZM_matriz = ZMCEM(theta=c(67,2),data = simdataaaa$simul_data[,,1])
 
 mu_estimate = MuMCEM(data = simdataaaa$simul_data[,,1], simZ = ZM_matriz)
 
 sigma_estimate = sigmaMCEM(data = simdataaaa$simul_data[,,1],
                            simZ = ZM_matriz,
                            mupd = mu_estimate)
 sigma_estimate
}
